% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testyverse-package.R
\docType{package}
\name{testyverse-package}
\alias{testyverse}
\alias{testyverse-package}
\title{testyverse: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Russ Hyde \email{russ.hyde.data@gmail.com}

}
\keyword{internal}
